move internal events & remove unused events
