+++ lib/src/rtc_engine.dart
@@ -9,6 +9,7 @@ import 'constants.dart';
 import 'events.dart';
 import 'exceptions.dart';
 import 'extensions.dart';
+import 'internal/events.dart';
 import 'logger.dart';
 import 'managers/delay.dart';
 import 'managers/event.dart';
